refactor(go-ai-client): split main.ts into libs/, fix lint and knip
