+++ knip.config.ts
@@ -25,6 +25,7 @@ const config: KnipConfig = {
     },
     // Lambda functions
     'functions/*': {
+      entry: ['src/handler.ts'],
       project: ['src/**/*.ts'],
     },
   },
@@ -40,6 +41,9 @@ const config: KnipConfig = {
     // ESLint plugins used in flat config (eslint.config.mjs)
     '@typescript-eslint/eslint-plugin',
     '@typescript-eslint/parser',
+    // Bundled into go-AILambda via esbuild — used transitively by @go-automation/go-ai
+    '@aws-sdk/client-bedrock-runtime',
+    '@aws-sdk/credential-provider-ini',
   ],
   ignoreBinaries: [
     // Used in CI security workflow
