+++ scripts/go/go-ai-client/package.json
@@ -20,7 +20,6 @@
   "author": "Team GO - Gestione Operativa",
   "license": "ISC",
   "dependencies": {
-    "@aws-sdk/client-bedrock-runtime": "^3.600.0",
     "@aws-sdk/client-lambda": "^3.600.0",
     "@aws-sdk/credential-provider-ini": "^3.600.0",
     "@go-automation/go-ai": "workspace:*",
